integrate local dependenct
