+++ example/package.json
@@ -10,7 +10,8 @@
     "build:tsc": "tsc",
     "start:prod": "node ./dist/app.js",
     "build": "npm-run-all build:clean build:tsc",
-    "start": "npm-run-all build start:prod",
+    "start": "npm-run-all build:lib build start:prod",
+    "build:lib": "cd .. && yarn build && cd example",
     "start:dev": "nodemon --watch src -e ts,ejs,json --exec npm run start",
     "clean": "rm -rf node_modules && rm -rf dist/* && yarn install",
     "codegen": "graphql-codegen --config codegen.yml"
@@ -24,7 +25,7 @@
     "body-parser": "^1.20.2",
     "express": "^4.18.2",
     "graphql": "^16.6.0",
-    "graphql-inherits": "^1.1.0",
+    "graphql-inherits": "../lib",
     "graphql-tools": "^8.3.20",
     "http": "^0.0.1-security",
     "npm-run-all": "^4.1.5",
